Specify library dependencies in library.properties
Specifying the library dependencies in the depends field of library.properties causes the Arduino Library Manager (Arduino IDE 1.8.10 and newer) to offer to install any missing dependencies during installation of this library. "arduino-cli lib install" will automatically install the dependencies (arduino-cli 0.7.0 and newer).

NOTE: I did not add the ArduinoJson dependency because the Arduino Library Manager dependency system is currently only capable of installing the newest version of each dependency and this library is not compatible with the 6.x.x versions of ArduinoJson. Once this library is updated to be compatible with the modern ArduinoJson library, that dependency should be added to the depends field as well.